backend/requirements.txt
# requirements.txt

# Core Flask
flask==2.0.3
werkzeug==2.0.3

# ASGI Server
gunicorn==20.1.0

# Celery
celery==5.2.7
redis==4.3.4

# Security
cryptography==36.0.1

# QR Generation
qrcode[pil]==7.3.1

# Environment
python-dotenv==0.19.2

# Thrift (pinned to avoid conflicts)
thriftpy2==0.5.2
